bot/requirements.txt
requests
loguru
pyTelegramBotApi
aiogram
BeautifulSoup4
python-dotenv